refactors to books to library
