+++ components/nav.data.ts
@@ -4,7 +4,7 @@ const menuItems: MenuItems = [
   { href: '/', text: 'Home', enabled: true },
   { href: '/posts', text: '📒 Posts', enabled: true },
   { href: '/experience', text: '💻 Experience', enabled: true },
-  { href: '/books', text: '📚 Books', enabled: true },
+  { href: '/library', text: '📚 Library', enabled: true },
 ].filter(menu =>  menu.enabled);
